RES-326: separated plotting from data generation
